spi: rtio: Move spi_rtio_copy to spi_rtio
To group all common APIs for SPI RTIO.
