+++ drivers/spi/spi_rtio.c
@@ -110,6 +110,164 @@ void spi_rtio_iodev_default_submit(const struct device *dev,
 }
 
 /**
+ * @brief Copy the tx_bufs and rx_bufs into a set of RTIO requests
+ *
+ * @param[in] r rtio context
+ * @param[in] iodev iodev to transceive with
+ * @param[in] tx_bufs transmit buffer set
+ * @param[in] rx_bufs receive buffer set
+ * @param[out] last_sqe last sqe submitted, NULL if not enough memory
+ *
+ * @retval Number of submission queue entries
+ * @retval -ENOMEM out of memory
+ */
+int spi_rtio_copy(struct rtio *r,
+		  struct rtio_iodev *iodev,
+		  const struct spi_buf_set *tx_bufs,
+		  const struct spi_buf_set *rx_bufs,
+		  struct rtio_sqe **last_sqe)
+{
+	int ret = 0;
+	size_t tx_count = tx_bufs ? tx_bufs->count : 0;
+	size_t rx_count = rx_bufs ? rx_bufs->count : 0;
+
+	uint32_t tx = 0, tx_len = 0;
+	uint32_t rx = 0, rx_len = 0;
+	uint8_t *tx_buf, *rx_buf;
+
+	struct rtio_sqe *sqe = NULL;
+
+	if (tx < tx_count) {
+		tx_buf = tx_bufs->buffers[tx].buf;
+		tx_len = tx_bufs->buffers[tx].len;
+	} else {
+		tx_buf = NULL;
+		tx_len = rx_bufs->buffers[rx].len;
+	}
+
+	if (rx < rx_count) {
+		rx_buf = rx_bufs->buffers[rx].buf;
+		rx_len = rx_bufs->buffers[rx].len;
+	} else {
+		rx_buf = NULL;
+		rx_len = tx_bufs->buffers[tx].len;
+	}
+
+
+	while ((tx < tx_count || rx < rx_count) && (tx_len > 0 || rx_len > 0)) {
+		sqe = rtio_sqe_acquire(r);
+
+		if (sqe == NULL) {
+			ret = -ENOMEM;
+			rtio_sqe_drop_all(r);
+			goto out;
+		}
+
+		ret++;
+
+		/* If tx/rx len are same, we can do a simple transceive */
+		if (tx_len == rx_len) {
+			if (tx_buf == NULL) {
+				rtio_sqe_prep_read(sqe, iodev, RTIO_PRIO_NORM,
+						   rx_buf, rx_len, NULL);
+			} else if (rx_buf == NULL) {
+				rtio_sqe_prep_write(sqe, iodev, RTIO_PRIO_NORM,
+						    tx_buf, tx_len, NULL);
+			} else {
+				rtio_sqe_prep_transceive(sqe, iodev, RTIO_PRIO_NORM,
+							 tx_buf, rx_buf, rx_len, NULL);
+			}
+			tx++;
+			rx++;
+			if (rx < rx_count) {
+				rx_buf = rx_bufs->buffers[rx].buf;
+				rx_len = rx_bufs->buffers[rx].len;
+			} else {
+				rx_buf = NULL;
+				rx_len = 0;
+			}
+			if (tx < tx_count) {
+				tx_buf = tx_bufs->buffers[tx].buf;
+				tx_len = tx_bufs->buffers[tx].len;
+			} else {
+				tx_buf = NULL;
+				tx_len = 0;
+			}
+		} else if (tx_len == 0) {
+			rtio_sqe_prep_read(sqe, iodev, RTIO_PRIO_NORM,
+					   (uint8_t *)rx_buf,
+					   (uint32_t)rx_len,
+					   NULL);
+			rx++;
+			if (rx < rx_count) {
+				rx_buf = rx_bufs->buffers[rx].buf;
+				rx_len = rx_bufs->buffers[rx].len;
+			} else {
+				rx_buf = NULL;
+				rx_len = 0;
+			}
+		} else if (rx_len == 0) {
+			rtio_sqe_prep_write(sqe, iodev, RTIO_PRIO_NORM,
+					    (uint8_t *)tx_buf,
+					    (uint32_t)tx_len,
+					    NULL);
+			tx++;
+			if (tx < tx_count) {
+				tx_buf = rx_bufs->buffers[rx].buf;
+				tx_len = rx_bufs->buffers[rx].len;
+			} else {
+				tx_buf = NULL;
+				tx_len = 0;
+			}
+		} else if (tx_len > rx_len) {
+			rtio_sqe_prep_transceive(sqe, iodev, RTIO_PRIO_NORM,
+						 (uint8_t *)tx_buf,
+						 (uint8_t *)rx_buf,
+						 (uint32_t)rx_len,
+						 NULL);
+			tx_len -= rx_len;
+			tx_buf += rx_len;
+			rx++;
+			if (rx < rx_count) {
+				rx_buf = rx_bufs->buffers[rx].buf;
+				rx_len = rx_bufs->buffers[rx].len;
+			} else {
+				rx_buf = NULL;
+				rx_len = tx_len;
+			}
+		} else if (rx_len > tx_len) {
+			rtio_sqe_prep_transceive(sqe, iodev, RTIO_PRIO_NORM,
+						 (uint8_t *)tx_buf,
+						 (uint8_t *)rx_buf,
+						 (uint32_t)tx_len,
+						 NULL);
+			rx_len -= tx_len;
+			rx_buf += tx_len;
+			tx++;
+			if (tx < tx_count) {
+				tx_buf = tx_bufs->buffers[tx].buf;
+				tx_len = tx_bufs->buffers[tx].len;
+			} else {
+				tx_buf = NULL;
+				tx_len = rx_len;
+			}
+		} else {
+			__ASSERT(false, "Invalid %s state", __func__);
+		}
+
+		sqe->flags = RTIO_SQE_TRANSACTION;
+	}
+
+	if (sqe != NULL) {
+		sqe->flags = 0;
+		*last_sqe = sqe;
+	}
+
+out:
+	return ret;
+}
+
+/**
  * @brief Lock the SPI RTIO spinlock
  *
  * This is used internally for controlling the SPI RTIO context, and is
